+++ __init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.1.2"
+__version__ = "1.1.3"
 
 from .pybettor import *  # noqa: F401, F403